test(web/services): create test file for orders service
